src/requirements.txt
enum34==1.1.6
futures==3.1.1
grpcio==1.7.0
grpcio-tools==1.7.0
protobuf==3.4.0
six==1.11.0
-e .
